requirements.txt
aiogram==3.4.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3